add types to babel-plugin-mdx-browser.ts
Based on my best inferences these are the closest types I could find.

TL;DR: The babel types seem set up in a fairly annoying way as I had to
end up setting up type guards that looked compatible.
If I'm doing anything dumb to prevent typescript from catching these
types correctly let me know.

---

Path is probably a `NodePath` from `@babel/traverse`

(Types from here on are from `@babel/types`)

`NodePath` takes a type `Node` and the only `Node` with `declarations`
I've found is `VariableDeclaration`. Also it matches the name.
(`ImportDeclaration` matches the other function but that one doesn't
really need it)

Then `declarations[0]` should be of type `VariableDeclarator` and it's `id` is
always `LVal` and nothing matching `LVal` has `name`.
For this reason I've removed this code.
If this code is actually useful then either my understanding so-far is
wrong or the types are wrong.

---

As mentioned previously `declarations[0]` should be of type `VariableDeclarator`
which is `Expression | null`.

The types compatible with `Expression` that has a `callee` are: `CallExpression`
`NewExpression` `BindExpression` `PipelineBareFunction` `OptionalCallExpression`.

Of these the only one I can find that is compatible with a type that has `name` is
`CallExpression` with `Expression | V8IntrinsicIdentifier`.